Commit de atualizacao do projeto

+++ src/main/java/com/projecttattoo/BrenoLendaTattoo/repositories/LoginsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projecttattoo.BrenoLendaTattoo.models.Logins;
 
-public interface LoginsRepository extends JpaRepository<Logins, Long>{
+public interface LoginsRepository extends JpaRepository<Logins, Integer>{
 	public Logins findByEmail(String email);
 	public Logins findByVerificationCode(String verificationCode);
 }
